refactor: moving the checkout step to an idependent component
