+++ src/app/quiz/page.tsx
@@ -3,28 +3,17 @@ import { motion } from 'framer-motion';
 import useQuiz from '@/hooks/useQuiz';
 import ProgressBar from '@/components/ProgressBar';
 import React, { useState } from 'react';
-import { useRouter } from 'next/navigation';
+import Checkout from '@/components/Checkout';
 
 export default function Quiz() {
   const [currentQuestionIndex, setCurrentQuestionIndex] = useState<number>(0);
   const [isCheckoutOpen, setIsCheckoutOpen] = useState<boolean>(false);
-  const router = useRouter();
   const { questions, answerQuestion } = useQuiz();
-
-  const [authorize, setAuthorize] = useState<boolean>(true);
 
   const currentQuestion = questions[currentQuestionIndex];
 
   const toggleCheckoutIsOpen = () => {
     setIsCheckoutOpen((c) => !c);
-  };
-
-  const handleGoToResults = () => {
-    router.push('/quiz/results');
-  };
-
-  const handleAuthorize = () => {
-    setAuthorize((authorize) => !authorize);
   };
 
   const handleNextQuestion = () => {
@@ -48,8 +37,6 @@ export default function Quiz() {
   */
 
   if (!questions.length) return null;
-
-  console.log(questions);
 
   return !isCheckoutOpen ? (
     <div className="h-full w-full flex flex-col items-center justify-center">
@@ -120,29 +107,8 @@ export default function Quiz() {
       </motion.div>
     </div>
   ) : (
-    <div className="h-full w-full flex flex-col items-center">
-      <div className="flex flex-col mt-6 h-fit py-5 rounded-xl my-2 w-[95%] bg-[#E7E7E7]">
-        <h1 className="font-bold text-2xl text-center">Checkout</h1>
-        <div className="flex flex-col my-2 px-5">
-          <label className="font-semibold" htmlFor="text">
-            Email
-          </label>
-          <input className="rounded-md h-[2rem]" type="text" />
-        </div>
-        <div className="flex flex-row my-2 px-5">
-          <input
-            type="checkbox"
-            checked={authorize}
-            onChange={handleAuthorize}
-          />
-          <p className="text-sm ml-1">Concedo meus dados para uso futuro</p>
-        </div>
-        <button
-          onClick={handleGoToResults}
-          className="p-6 mt-5 mx-5 rounded-xl bg-[#009B92] text-white font-bold text-center">
-          Seguir para resultados
-        </button>
-      </div>
-    </div>
+    <>
+      <Checkout />
+    </>
   );
 }
